CI: fix all ruff errors + relax stylistic rules
The CI workflow was firing failure emails on every push because ruff
had 42 errors against the codebase — most stylistic noise, some real.

Ruff config: ignore six stylistic rules that are pure noise for this
project's style:
- B904   (raise X from err) — many re-raises clean the message
- B008   (function-call-in-default) — typer.Option/Argument require it
- RUF002 / RUF003 — em-dashes in docstrings/comments are intentional
- RUF006 — fire-and-forget asyncio tasks are intentional

Auto-fixed by ruff --fix --unsafe-fixes (17 errors):
- I001 unsorted imports → reorganised
- F401 unused imports → removed
- RUF019 unnecessary key check → simplified
- RUF021 chained and/or precedence → parenthesised
- RUF046 unnecessary int cast → removed
- RUF100 unused noqa → cleared
- UP037 quoted annotations → unquoted

Manual fix (5 errors):
- src/mlstudio/api/server.py:627-629 — three E701 (multi-statement
  colons in a tight if/elif chain inside stats build) split onto
  their own lines.
- src/mlstudio/profiles/mst.py:20 — unused enumerate index in
  build_mst's node-adding loop → drop enumerate().
- src/mlstudio/schemes/cgmlst_org.py:77 — E702 semicolon-joined
  buf.write + total += in the streaming download → split.

Tests still pass (7/7); ruff is now clean.

+++ src/mlstudio/profiles/distance.py
@@ -82,7 +82,7 @@ def hamming_matrix(
                     if a != b:
                         d += 1
             if policy == "scaled" and shared > 0 and n_loci > 0:
-                d = int(round(d * n_loci / shared))
+                d = round(d * n_loci / shared)
             if policy == "pairwise_complete" and shared == 0:
                 d = n_loci if n_loci > 0 else 1
             mat[i, j] = d
